modif simon (cast d'un char en int dans l'affichage du level)

+++ client/src/GUI/GUI.cpp
@@ -45,13 +45,11 @@ GUI::~GUI()
 void		GUI::callback()
 {
 	Clock timeInCallback;
-//#ifndef _WIN32
 	// limitation
 	if (timerLastCallback.getTimeMilli() < 1000 / FPS_MAX)
 		return ;
 	timerLastCallback.reset();
 	// fps
-//#endif // !WIN_32
 
 	fps++;
 	if (timerFps.getTimeMilli() >= 1000)
@@ -1267,7 +1265,7 @@ void	GUI::setStagePopup(uint8_t nStage)
 
 	// Show Popup
 	std::stringstream ss;
-	ss << "Level : " << nStage << " !";
+	ss << "Level : " << (int) nStage << " !";
 	_levelWidget->showPopup(ss.str(), 3);
 
 	// Define position
